Renamed PresenterBinder -> LifecycleBinder
